Make sure a valid escape sequence is always returned
If for whatever reason the make-*-cursor-shape functions receive
something that's not 'box, 'bar, or 'hbar, currently they'll either
return nil and do nothing or trigger a void-variable error, causing
the user to have to read the code to see what's wrong.

This is mainly caused by insufficient let-binding and not providing a
fallthrough case for `cond`. Here I opted for the minimum fix, to just
make them assume their input is 'box if it's invalid.

This should fix #19.